Generate Text for Labels

This function combines texts (strings) or numbers specified in the Arguments of the function. Function is originally made to automate label generation for lab use.

Usage

Label_Gen(
  study_id,
  timepoint,
  trt_tank_id,
  animal,
  animal_numbers,
  tissue,
  repli_num,
  n_col = 6
)

Arguments

study_id

A single text (or vector). Must not be NA.

timepoint

A single text (or vector) representing the sampling timepoint(s). Set to = NA if not applicable.

trt_tank_id

A single text (or vector) representing the treatment and tank ID combinations. Set to = NA if not applicable.

animal

A single text (or vector) representing the group of animal(s). Set to = NA if not applicable.

animal_numbers

A numeric (or vector) representing the number IDs for the animals. Set to = NA if not applicable.

tissue

A single text (or vector) representing the tissue type(s). Set to = NA if not applicable.

repli_num

A numeric (or vector) representing the replicate number(s). Set to = NA if not applicable.

n_col

A numeric corresponding to the number of columns in the label paper (and in the output matrix). Defaults to 6.

Value

Returns a matrix of text representing the combinations. Matrix automatically saved in a .csv in your working directory.

Details

Computes all possible combination of variables provided in Arguments. Output combinations are sorted in an ascending alphanumeric order.

See also

Examples

# set working directory
setwd("C:/Users/sean4/Downloads") # make sure to use "/" instead of "\" (which is the default you might get from Windows)
#> Error in setwd("C:/Users/sean4/Downloads"): cannot change working directory

# apply Label_Gen()!
Label_Gen(study_id = "ONDA01180",
          timepoint = "Baseline",
          trt_tank_id = c("Tk.F01-TrtA", "Tk.F02-TrtA", "Tk.F03-TrtB", "Tk.F04-TrtB"),
          animal = "Fish",
          animal_numbers = 1:4,
          tissue = NA,
          repli_num = 1:2,
          n_col = 6)
#> Error in FUN(X[[i]], ...): object 'study_id' not found