Create Flextables from Prevalence Data

Converts the pathology data summary from Path_Summary() into a list of flextable objects to print out to Word. Summary data are converted to flextables in wide and long formats. Contrast data (pairwise comparison p-values) are converted to their standalone flextable.

Usage

Path_Table(path_sum, digits = 0, prev_append = c("se_plus-minus"))

Arguments

path_sum

A list of dataframes or a single dataframe from Path_Summary().

digits

A number specifying the decimal places for the prevalence values in the output flextable. Defaults to 0.

prev_append

A string specifying the suffix to append to prevalence values in the output table. Suffix may be the standard error or counts of presence. Suffix may be presented inside brackets or after a +/- sign. Choose one of the following arguments: "se_brackets", "se_plus-minus", "presence_brackets". Defaults to "se_plus-minus".

Value

A list that contains, at minimum, 2 flextable objects of the summary statistics (from Path_Summary()) in wide and long format. If a contrast dataframe from Path_Summary() is also included in the path_sum argument, then the output list contains additionally a flextable for the contrast dataframe.

Examples

#Generate summarized pathology dataframe:
data(path_db_ex)
path_db = Path_Gen(path_db = path_db_ex,
                   rel_cols = 1,
                   path_cols = 3:12)

path_sum = Path_Summary(path_db = path_db,
                        factor = c("Trt.ID"),
                        path_cols = 2:11,
                        contrast_out = TRUE)
#> [1] "Ascites"
#> [1] "Dark body"
#> [1] "Exopthlamia"
#> [1] "Eye haemorrhage"
#> [1] "Grey kidney"
#> [1] "Mouth haemorrhage"
#> [1] "Skin haemorrhage"
#> [1] "Swollen kidney"
#> [1] "Swollen spleen"
#> [1] "Visceral haemorrhage"

path_tables = Path_Table(path_sum = path_sum,
                         digits = 1, #1 decimal places for prevalence values
                         prev_append = c("se_brackets")) #append standard errors (in brackets) to prevalence values

#View pathology table (wide format):
path_tables$wide

Table 1. Pathological sign prevalence (%) and standard errors (in brackets) across treatments.

Trt.ID

Ascites

Dark body

Exopthlamia

Eye haemorrhage

Grey kidney

Mouth haemorrhage

Skin haemorrhage

Swollen kidney

Swollen spleen

Visceral haemorrhage

A

11.1 (6.0)

0.0 (0.0)

3.4 (3.4)

24.1 (7.9)

63.0 (9.3)a

72.4 (8.3)

41.4 (9.1)

88.9 (6.0)

55.6 (9.6)

70.4 (8.8)

B

23.1 (8.3)

0.0 (0.0)

0.0 (0.0)

14.3 (6.6)

80.8 (7.7)ab

50.0 (9.4)

35.7 (9.1)

96.2 (3.8)

46.2 (9.8)

57.7 (9.7)

C

24.0 (8.5)

0.0 (0.0)

0.0 (0.0)

21.4 (7.8)

60.0 (9.8)a

75.0 (8.2)

46.4 (9.4)

92.0 (5.4)

60.0 (9.8)

60.0 (9.8)

D

15.4 (7.1)

3.6 (3.5)

0.0 (0.0)

25.0 (8.2)

88.5 (6.3)ab

64.3 (9.1)

42.9 (9.4)

88.5 (6.3)

42.3 (9.7)

76.9 (8.3)

E

7.4 (5.0)

3.6 (3.5)

0.0 (0.0)

17.9 (7.2)

96.3 (3.6)b

60.7 (9.2)

46.4 (9.4)

88.9 (6.0)

66.7 (9.1)

74.1 (8.4)

Global-P

0.3718

0.4427

0.5264

0.8409

0.0023

0.2995

0.9239

0.8184

0.3737

0.4818

For every pathological sign, treatments were compared using logistic regression and likelihood ratio tests. Letters are based on pairwise comparisons of Trt.ID with p-values adjusted using the Benjamini-Hochberg method.

#View pathology table (long format):
path_tables$long

Table 2. Pathological signs prevalence (%) and standard errors (in brackets) across treatments.

Pathology

Trt.ID

Prevalence

n

Global-P
(significance of treatment)

Letters

Ascites

A

11.1 (6.0)

27

0.3718

a

B

23.1 (8.3)

26

a

C

24.0 (8.5)

25

a

D

15.4 (7.1)

26

a

E

7.4 (5.0)

27

a

Dark body

A

0.0 (0.0)

29

0.4427

a

B

0.0 (0.0)

28

a

C

0.0 (0.0)

28

a

D

3.6 (3.5)

28

a

E

3.6 (3.5)

28

a

Exopthlamia

A

3.4 (3.4)

29

0.5264

a

B

0.0 (0.0)

28

a

C

0.0 (0.0)

28

a

D

0.0 (0.0)

28

a

E

0.0 (0.0)

28

a

Eye haemorrhage

A

24.1 (7.9)

29

0.8409

a

B

14.3 (6.6)

28

a

C

21.4 (7.8)

28

a

D

25.0 (8.2)

28

a

E

17.9 (7.2)

28

a

Grey kidney

A

63.0 (9.3)

27

0.0023

a

B

80.8 (7.7)

26

ab

C

60.0 (9.8)

25

a

D

88.5 (6.3)

26

ab

E

96.3 (3.6)

27

b

Mouth haemorrhage

A

72.4 (8.3)

29

0.2995

a

B

50.0 (9.4)

28

a

C

75.0 (8.2)

28

a

D

64.3 (9.1)

28

a

E

60.7 (9.2)

28

a

Skin haemorrhage

A

41.4 (9.1)

29

0.9239

a

B

35.7 (9.1)

28

a

C

46.4 (9.4)

28

a

D

42.9 (9.4)

28

a

E

46.4 (9.4)

28

a

Swollen kidney

A

88.9 (6.0)

27

0.8184

a

B

96.2 (3.8)

26

a

C

92.0 (5.4)

25

a

D

88.5 (6.3)

26

a

E

88.9 (6.0)

27

a

Swollen spleen

A

55.6 (9.6)

27

0.3737

a

B

46.2 (9.8)

26

a

C

60.0 (9.8)

25

a

D

42.3 (9.7)

26

a

E

66.7 (9.1)

27

a

Visceral haemorrhage

A

70.4 (8.8)

27

0.4818

a

B

57.7 (9.7)

26

a

C

60.0 (9.8)

25

a

D

76.9 (8.3)

26

a

E

74.1 (8.4)

27

a

For every pathological sign, treatments were compared pairwise using logistic regression and likelihood ratio tests. Letters are based on pairwise comparisons of Trt.ID with p-values adjusted using the Benjamini-Hochberg method.

#View pathology contrast table:
path_tables$contrast

Table 3. Comparison of pathological sign prevalence across pairs of treatments.

Pathology

Compared pair

P-value

Ascites

A vs. B

0.6068

A vs. C

0.6068

A vs. D

0.7174

A vs. E

0.7174

B vs. C

0.9381

B vs. D

0.6863

B vs. E

0.5240

C vs. D

0.6863

C vs. E

0.5240

D vs. E

0.6863

Dark
body

A vs. B

1.0000

A vs. C

1.0000

A vs. D

0.3933

A vs. E

0.3933

B vs. C

1.0000

B vs. D

0.3933

B vs. E

0.3933

C vs. D

0.3933

C vs. E

0.3933

D vs. E

1.0000

Exopthlamia

A vs. B

0.6051

A vs. C

0.6051

A vs. D

0.6051

A vs. E

0.6051

B vs. C

1.0000

B vs. D

1.0000

B vs. E

1.0000

C vs. D

1.0000

C vs. E

1.0000

D vs. E

1.0000

Eye
haemorrhage

A vs. B

0.8971

A vs. C

0.8971

A vs. D

0.9397

A vs. E

0.8971

B vs. C

0.8971

B vs. D

0.8971

B vs. E

0.8971

C vs. D

0.8971

C vs. E

0.8971

D vs. E

0.8971

Grey
kidney

A vs. B

0.2102

A vs. C

0.8263

A vs. D

0.0687

A vs. E

0.0061

B vs. C

0.1689

B vs. D

0.4889

B vs. E

0.1284

C vs. D

0.0581

C vs. E

0.0061

D vs. E

0.3388

Mouth
haemorrhage

A vs. B

0.4039

A vs. C

0.8245

A vs. D

0.6362

A vs. E

0.5993

B vs. C

0.4039

B vs. D

0.5993

B vs. E

0.5993

C vs. D

0.5993

C vs. E

0.5993

D vs. E

0.8245

Skin
haemorrhage

A vs. B

0.9851

A vs. C

0.9851

A vs. D

1.0000

A vs. E

0.9851

B vs. C

0.9851

B vs. D

0.9851

B vs. E

0.9851

C vs. D

0.9851

C vs. E

1.0000

D vs. E

0.9851

Swollen
kidney

A vs. B

1.0000

A vs. C

1.0000

A vs. D

1.0000

A vs. E

1.0000

B vs. C

1.0000

B vs. D

1.0000

B vs. E

1.0000

C vs. D

1.0000

C vs. E

1.0000

D vs. E

1.0000

Swollen
spleen

A vs. B

0.7048

A vs. C

0.7801

A vs. D

0.6682

A vs. E

0.6695

B vs. C

0.6682

B vs. D

0.7801

B vs. E

0.6532

C vs. D

0.6682

C vs. E

0.7724

D vs. E

0.6532

Visceral
haemorrhage

A vs. B

0.6705

A vs. C

0.7200

A vs. D

0.8400

A vs. E

0.8670

B vs. C

0.8670

B vs. D

0.6705

B vs. E

0.6705

C vs. D

0.6705

C vs. E

0.6705

D vs. E

0.8670

Data for every pathological sign was fitted to a logistic regression model used to calculate P values. Pairwise comparison of treatments were conducted using likelihood ratio tests. P-values from pairwise comparison were adjusted using the Benjamini-Hochberg method. Different letters denote significantly different groups.

#Print results in word docx:
library(officer)
library(flextable)
results_doc = read_docx() |>
  body_add_flextable(value = path_tables$wide, align = "left", topcaption = TRUE) |>
  body_add_break() |>
  body_add_flextable(value = path_tables$long, align = "left", topcaption = TRUE) |>
  body_add_break() |>
  body_add_flextable(value = path_tables$contrast, align = "left", topcaption = TRUE)
print(results_doc, target = paste0("ONDA Pathology Analyses ", format(Sys.Date(), "%d%b%Y"),".docx"))