Generate Survivor Data

Produces survival data that includes rows for every surviving fish based on the starting number of fish and mortality data. To generate survivor data for tanks absent in the input mortality dataframe, specify the arguments tank_without_mort and trt_without_mort. To generate survivor data with tank specific starting numbers of fish, input a dataframe into the argument starting_fish_count instead of a single value; details in Arguments.

Usage

Surv_Gen(
  mort_db,
  starting_fish_count,
  last_tte,
  tank_without_mort = NULL,
  trt_without_mort = NULL
)

Arguments

mort_db

A mort dataframe as described in Details.

starting_fish_count

Value representing the starting number of fish for every tank. Alternatively, a dataframe containing the columns "Trt.ID", "Tank.ID", and "starting_fish_count" to allow for different fish starting numbers per tank.

last_tte

Value representing the time-to-event the fish survived to, assigned to every row of survivor data generated.

tank_without_mort

A vector of strings specifying the tanks absent from mort_db; used to generate survivor data for those tanks. Argument ignored if starting_fish_count is a dataframe.

trt_without_mort

A vector of strings corresponding to tank_without_mort. Keep their order the same. Argument ignored if starting_fish_count is a dataframe.

Value

A dataframe produced by combining the input mort data and generated rows of survivor data.

Details

The mort dataframe supplied as input should consist of the following 4 columns at minimum:

"Trt.ID" = Labels for treatment groups in the study.

"Tank.ID" = Labels for tanks in the study (each tank must have a unique label).

"TTE" = Time to Event. Event could be fish death or being sampled and removed depending on "Status".

"Status" = Value indicating what happened at TTE. 1 for dead fish, 0 for those sampled and removed.

Each row should represent one fish.

For an example dataframe, execute data(mort_db_ex) and view.

Examples

#First, we load an example mortality database available from the safuncs package
data(mort_db_ex)

#Next, we input this data into Surv_Gen() as well as the study details to generate entries (rows) for survivors in the output - a "complete" dataframe for further survival analysis and data visualization.
Surv_Gen(mort_db = mort_db_ex,
         starting_fish_count = 100,
         last_tte = 54,
         tank_without_mort = c("C99", "C100"),
         trt_without_mort = c("A", "B"))
#>     Tank.ID Trt.ID TTE Status
#> 1        C1      B   0      1
#> 2        C6      B   4      1
#> 3        C6      B   7      1
#> 4        C1      B  19      1
#> 5        C8      D  19      1
#> 6       C11      B  20      1
#> 7       C11      B  20      1
#> 8       C12      C  20      1
#> 9        C1      B  21      1
#> 10       C8      D  21      1
#> 11       C1      B  22      1
#> 12       C3      C  22      1
#> 13       C6      B  22      1
#> 14       C8      D  22      1
#> 15       C9      A  23      1
#> 16       C1      B  24      1
#> 17       C3      C  24      1
#> 18      C11      B  24      1
#> 19       C6      B  25      1
#> 20       C8      D  25      1
#> 21      C11      B  25      1
#> 22       C2      A  26      1
#> 23       C3      C  26      1
#> 24       C5      D  26      1
#> 25       C5      D  26      1
#> 26       C9      A  26      1
#> 27       C9      A  26      1
#> 28      C11      B  26      1
#> 29      C12      C  26      1
#> 30       C9      A  26      1
#> 31      C11      B  26      1
#> 32      C11      B  26      1
#> 33      C11      B  26      1
#> 34       C9      A  26      1
#> 35       C1      B  27      1
#> 36       C3      C  27      1
#> 37       C4      D  27      1
#> 38       C5      D  27      1
#> 39       C6      B  27      1
#> 40       C6      B  27      1
#> 41      C11      B  27      1
#> 42      C11      B  27      1
#> 43      C11      B  27      1
#> 44      C11      B  27      1
#> 45      C14      A  27      1
#> 46      C15      C  27      1
#> 47       C2      A  27      1
#> 48       C2      A  27      1
#> 49       C3      C  27      1
#> 50       C1      B  27      1
#> 51       C9      A  27      1
#> 52       C6      B  28      1
#> 53       C6      B  28      1
#> 54       C8      D  28      1
#> 55      C11      B  28      1
#> 56      C11      B  28      1
#> 57      C12      C  28      1
#> 58      C15      C  28      1
#> 59       C3      C  28      1
#> 60       C1      B  29      1
#> 61       C4      D  29      1
#> 62       C4      D  29      1
#> 63       C5      D  29      1
#> 64       C5      D  29      1
#> 65       C6      B  29      1
#> 66       C6      B  29      1
#> 67       C6      B  29      1
#> 68       C9      A  29      1
#> 69      C11      B  29      1
#> 70      C11      B  29      1
#> 71      C11      B  29      1
#> 72      C12      C  29      1
#> 73       C5      D  29      1
#> 74       C5      D  29      1
#> 75       C5      D  29      1
#> 76      C11      B  29      1
#> 77      C15      C  29      1
#> 78       C1      B  30      1
#> 79       C1      B  30      1
#> 80       C1      B  30      1
#> 81       C1      B  30      1
#> 82       C2      A  30      1
#> 83       C2      A  30      1
#> 84       C3      C  30      1
#> 85       C4      D  30      1
#> 86       C4      D  30      1
#> 87       C4      D  30      1
#> 88       C4      D  30      1
#> 89       C4      D  30      1
#> 90       C5      D  30      1
#> 91       C6      B  30      1
#> 92       C6      B  30      1
#> 93       C6      B  30      1
#> 94       C6      B  30      1
#> 95       C6      B  30      1
#> 96       C6      B  30      1
#> 97       C8      D  30      1
#> 98      C11      B  30      1
#> 99      C11      B  30      1
#> 100     C11      B  30      1
#> 101     C11      B  30      1
#> 102     C12      C  30      1
#> 103     C12      C  30      1
#> 104     C12      C  30      1
#> 105     C15      C  30      1
#> 106      C1      B  30      1
#> 107     C14      A  30      1
#> 108      C2      A  31      1
#> 109      C2      A  31      1
#> 110      C3      C  31      1
#> 111      C3      C  31      1
#> 112      C4      D  31      1
#> 113      C4      D  31      1
#> 114      C4      D  31      1
#> 115      C4      D  31      1
#> 116      C6      B  31      1
#> 117      C6      B  31      1
#> 118      C8      D  31      1
#> 119      C8      D  31      1
#> 120      C8      D  31      1
#> 121      C8      D  31      1
#> 122      C8      D  31      1
#> 123      C9      A  31      1
#> 124     C11      B  31      1
#> 125     C11      B  31      1
#> 126     C12      C  31      1
#> 127     C12      C  31      1
#> 128     C12      C  31      1
#> 129     C14      A  31      1
#> 130     C15      C  31      1
#> 131      C1      B  31      1
#> 132      C3      C  31      1
#> 133      C4      D  31      1
#> 134     C11      B  31      1
#> 135      C2      A  32      1
#> 136      C2      A  32      1
#> 137      C2      A  32      1
#> 138      C2      A  32      1
#> 139      C2      A  32      1
#> 140      C3      C  32      1
#> 141      C4      D  32      1
#> 142      C4      D  32      1
#> 143      C6      B  32      1
#> 144      C9      A  32      1
#> 145     C11      B  32      1
#> 146     C12      C  32      1
#> 147     C14      A  32      1
#> 148      C2      A  32      1
#> 149      C4      D  32      1
#> 150      C4      D  32      1
#> 151      C6      B  32      1
#> 152      C1      B  33      1
#> 153      C1      B  33      1
#> 154      C1      B  33      1
#> 155      C1      B  33      1
#> 156      C2      A  33      1
#> 157      C2      A  33      1
#> 158      C3      C  33      1
#> 159      C3      C  33      1
#> 160      C3      C  33      1
#> 161      C3      C  33      1
#> 162      C3      C  33      1
#> 163      C4      D  33      1
#> 164      C5      D  33      1
#> 165      C6      B  33      1
#> 166      C6      B  33      1
#> 167      C6      B  33      1
#> 168      C8      D  33      1
#> 169      C9      A  33      1
#> 170     C11      B  33      1
#> 171     C11      B  33      1
#> 172     C12      C  33      1
#> 173     C12      C  33      1
#> 174     C12      C  33      1
#> 175     C12      C  33      1
#> 176     C14      A  33      1
#> 177     C15      C  33      1
#> 178     C15      C  33      1
#> 179      C2      A  33      1
#> 180      C2      A  33      1
#> 181      C1      B  34      1
#> 182      C1      B  34      1
#> 183      C1      B  34      1
#> 184      C2      A  34      1
#> 185      C2      A  34      1
#> 186      C2      A  34      1
#> 187      C3      C  34      1
#> 188      C3      C  34      1
#> 189      C3      C  34      1
#> 190      C3      C  34      1
#> 191      C3      C  34      1
#> 192      C5      D  34      1
#> 193      C5      D  34      1
#> 194      C5      D  34      1
#> 195      C5      D  34      1
#> 196      C6      B  34      1
#> 197      C6      B  34      1
#> 198      C8      D  34      1
#> 199      C8      D  34      1
#> 200      C9      A  34      1
#> 201      C9      A  34      1
#> 202     C11      B  34      1
#> 203     C12      C  34      1
#> 204     C12      C  34      1
#> 205     C14      A  34      1
#> 206     C15      C  34      1
#> 207      C8      D  34      1
#> 208      C1      B  35      1
#> 209      C1      B  35      1
#> 210      C1      B  35      1
#> 211      C2      A  35      1
#> 212      C2      A  35      1
#> 213      C2      A  35      1
#> 214      C2      A  35      1
#> 215      C2      A  35      1
#> 216      C3      C  35      1
#> 217      C3      C  35      1
#> 218      C3      C  35      1
#> 219      C4      D  35      1
#> 220      C4      D  35      1
#> 221      C4      D  35      1
#> 222      C5      D  35      1
#> 223      C5      D  35      1
#> 224      C5      D  35      1
#> 225      C5      D  35      1
#> 226      C5      D  35      1
#> 227      C6      B  35      1
#> 228      C6      B  35      1
#> 229      C6      B  35      1
#> 230      C8      D  35      1
#> 231      C8      D  35      1
#> 232      C8      D  35      1
#> 233      C8      D  35      1
#> 234      C8      D  35      1
#> 235      C9      A  35      1
#> 236      C9      A  35      1
#> 237     C11      B  35      1
#> 238     C11      B  35      1
#> 239     C12      C  35      1
#> 240     C12      C  35      1
#> 241     C14      A  35      1
#> 242     C15      C  35      1
#> 243     C15      C  35      1
#> 244     C15      C  35      1
#> 245     C15      C  35      1
#> 246     C15      C  35      1
#> 247      C1      B  36      1
#> 248      C1      B  36      1
#> 249      C2      A  36      1
#> 250      C2      A  36      1
#>  [ reached 'max' / getOption("max.print") -- omitted 1150 rows ]