Calculates Power for Survival Experiments

Calculates the power of global and/or pairwise hypothesis tests for survival studies with support over a range of experimental designs. This versatility is enabled by the simulation-based approach of the power calculation, using the modular Surv_Simul() to simulate survival data of various experimental designs. Power calculations can be made to account for inter-tank variation using a mixed cox proportional hazards model (set argument model = "coxph_glmm"). Additionally, power calculations can account for the multiplicity of pairwise comparisons using pairwise_corr. Users can compare power across different experimental designs by specifying each as a list element in Surv_Simul(). A brief tutorial is written in Examples (unfinished) to guide the user on how to use Surv_Power() to calculate power for various cases.

Usage

Surv_Power(
  simul_db = simul_db_ex,
  global_test = "logrank",
  model = NULL,
  pairwise_test = "logrank",
  pairwise_corr = "BH",
  prog_show = TRUE,
  plot_out = TRUE,
  plot_lines = FALSE,
  xlab = "List Element #",
  xnames = NULL,
  plot_save = TRUE
)

Arguments

simul_db

An output from Surv_Simul() which includes the survival dataframe simulated with the desired experimental design parameters.

global_test

A character vector representing the method(s) to use for global hypothesis testing of significance of treatment. Methods available are: "logrank", "wald", "score", "LRT". "logrank" represents the global logrank test of significance. The latter three methods are standard global hypothesis testing methods for models. They are only available when the argument model is specified (i.e. not NULL). "wald" represents the Wald Chisquare Test which assesses whether model parameters (log(hazard ratios)) jointly are significantly different from 0 (i.e. HRs ≠ 1). Wald test can be done for various cox-proportional hazard models that could be relevant to our studies (glm, glmm, and gee). Due to its broad applicability, while also producing practically the same p-value most of the time compared to the other model tests, "wald" is the recommended option of the three. "score" represents the Lagrange multiplier or Score test. 'LRT' represents the likelihood ratio test. Defaults to "logrank" for now due to its ubiquity of use.

model

A character vector representing the model(s) to fit for hypothesis testing. Models available are: "coxph_glm" and "coxph_glmm". "coxph_glm" represents the standard cox proportional hazard model fitted using survival::coxph() with Trt.ID as a fixed factor. "coxph_glmm" represents the mixed cox proportional hazard model fitted using coxme::coxme() with Trt.ID as a fixed factor and Tank.ID as a random factor to account for inter-tank variation. Defaults to NULL where no model is fitted for hypothesis testing.

pairwise_test

A character vector representing the method(s) used for pairwise hypothesis tests. Use "logrank" to calculate power for logrank tests comparing different treatments. Use "EMM" to calculate power using Estimated Marginal Means based on model estimates (from 'coxph_glm' and/or 'coxph_glmm'). Defaults to "logrank".

pairwise_corr

A character vector representing the method(s) used to adjust p-values for multiplicity of pairwise comparisons. For clarification, this affects the power of the pairwise comparisons. Methods available are: "tukey", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", and "none". Under Details (yet to be finished), I discuss the common categories of adjustment methods and provided a recommendation for "BH". Defaults to "none" for now.

prog_show

Whether to display the progress of Surv_Power() by printing the number of sample sets with p-values calculated. Defaults to TRUE.

plot_out

Whether to output plots illustrating the power of global and/or pairwise hypothesis tests. Defaults to TRUE.

plot_lines

Whether to plot lines connecting points of the same hypothesis test in the plot output. Defaults to TRUE.

xlab

A string representing the x-axis title. Defaults to "List Element #".

xnames

A character vector of names for x-axis labels. Defaults to NULL where names are the list element numbers from Surv_Simul().

plot_save

Whether to save plots as a .tiff in the working directory. Defaults to TRUE.

Value

Outputs a list containing any of the following four items depending on input arguments:

When global_test ≠ NULL, outputs a dataframe named power_glob_db containing power values calculated for global hypothesis tests. The dataframe consists of six columns:

model		The type of model being evaluated in power calculations

global_test		The global hypothesis test being evaluated

power		The percentage of p-values below 0.05, i.e. power

power_se		Standard error for percentages

sample_sets_n		Number of sample sets used in calculating power

list_element_num		The list element number associated with the power value calculated

When global_test ≠ NULL and plot_out = TRUE, a plot showing power values for global hypothesis test. Plot corresponds to power_glob_db.

When pairwise_test ≠ NULL, outputs a dataframe named power_pair_db containing power values for pairwise hypothesis tests. The dataframe consists of eight columns:

pair		The treatment groups to be compared

model		The type of model being evaluated in power calculations

pairwise_test		The type of pairwise_test being evaluated

pairwise_corr		The method of p-value correction/adjustments for paiwrise comparisons

power		The percentage of p-values below 0.05, i.e. power

power_se		Standard error for percentages

sample_sets_n		Number of sample sets used in calculating power

list_element_num		The list element number associated with the power value calculated

When pairwise_test ≠ NULL and plot_out = TRUE, a plot showing power values across list elements. Plot corresponds to power_pair_db.

See also

Examples

# Below is a tutorial (under construction) on how to calculate power using
# Surv_Power(). First, we'll have to simulate the future survival dataset using
# Surv_Simul().

# First, we'll load up a past data to use as a reference for the simulations.

haz_db_ex = data(haz_db_ex)
haz_db_ex = haz_db_ex[haz_db_ex$Trt.ID == "A"] # filter for control fish
#> Error in haz_db_ex$Trt.ID: $ operator is invalid for atomic vectors
head(haz_db_ex, n = 5)
#> [1] "haz_db_ex"

# As may be clear from the above, we are using the past data's hazard curve properties.
# NOTE: Whether the shape of the hazard curve repeats in the future study only matters
# to power calculation accuracy when the future study involves fish dropping out mid
# study (due to sampling for example).

# We will now plug in the past data into Surv_Simul() with the desired future
# experimental design parameters!

Surv_Simul()
#> Error in Surv_Simul(): argument "haz_db" is missing, with no default

# Next, we input the simulated data to Surv_Power() to calculate power!

# In the above example, power is calculated for a global logrank test and _.
# For example, based on model estimates, or to account for multiplicity of pairwise
# comparison as shown below:

# Additionally, we can investigate power across different experimental setups. What
# if we want to know the effects of a weaker or stronger challenge to
# power, for example. We can set this up using list_elements in Surv_Simul():

# What if we want to know if the sample size we need to get 80% power in the weak
# challenge scenario?

# The inter-tank variation ....Below are the results...
# under two different challenge strength scenarios to show the complexities of the
# influence of tank variation to power.

# Additionally, showcase multiple analysis tools in case (for e.g. correction methods
# and model fits..)
#