Predict Future Survival

Compare survival characteristics to a reference data and predict future survival. Comparisons shown using Kaplan-Meier Survival Plot and Hazard Time Plot which are saved automatically to the working directory. Predictions tabulated and its history visualized. Outputs further detailed in Value.

Usage

Surv_Pred(
  surv_db,
  ref_surv_db,
  pred_tte = NULL,
  pred_method = "simple",
  pred_dailybin = TRUE,
  pred_phi = 1.5,
  pred_lambda = NULL,
  plot_save = TRUE,
  plot_prefix = "ONDA_XX"
)

Arguments

surv_db

A survival dataframe as described in Surv_Plots(), consisting of the four columns named TTE, Status, Trt.ID and Tank.ID. Example: surv_db_ex which is generated using Surv_Gen().

ref_surv_db

A survival dataframe with the same column names as the supplied surv_db. Must only have one Trt.ID which represents the reference used for prediction.

pred_tte

A numeric representing the time to event at which the survival rate is to be predicted. Defaults to NULL where the maximum TTE observed in ref_surv_db is used.

pred_method

A string representing the method used to predict survival rate. Options are "simple" or "adaptive". Methods further discussed in Details. Defaults to "simple".

pred_dailybin

Argument ignored if ref_surv_db contains multiple tanks. Whether to use daily (1 TTE interval) time bins in estimating the reference hazard curve. Defaults to TRUE. Further details in Surv_Plots.

pred_phi

Argument ignored if ref_surv_db contains multiple tanks. A numeric indicating the dispersion parameter to be used in the count model that estimates the reference hazard curve. Defaults to 1.5. Set to NULL for a data driven estimate. Further details in Surv_Plots.

pred_lambda

Argument ignored if ref_surv_db contains multiple tanks. A numeric indicating the smoothing parameter to be used in estimating the reference hazard curve. Ignored if pred_phi is not NULL. Defaults to NULL (data driven estimate). Further details in Surv_Plots.

plot_save

Whether to save plot outputs in the working directory. Defaults to TRUE.

plot_prefix

A string specifying the prefix of the filename of the saved plots. Defaults to "ONDA_XX".

Value

Placeholder

Details

Placeholder

See also

Examples

#Placeholder