Predict End Survival Rate

Predict survival rate for a given survival dataset provided a reference survival database used to estimate a reference hazard curve. Prediction done separately by treatment group.

Usage

Surv_Pred(
  pred_db,
  ref_db,
  predsr_tte,
  method = 2,
  coxph_mod = "GLMM",
  lambda_pred = NULL,
  phi_pred = NULL
)

Arguments

pred_db

Placeholder

ref_db

Placeholder

predsr_tte

Placeholder

method

Placeholder

coxph_mod

Placeholder

lambda_pred

Placeholder

phi_pred

Placeholder

Value

Placeholder

Details

P

See also

Examples

#Placeholder